Document all new constructs
